Move API to service folder
